+++ src/DataTransferObjects/Price.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Cyberfusion\Oxxa\DataTransferObjects;
+
+class Price
+{
+    public function __construct(
+        public int $period,
+        public float $price,
+    ) {
+    }
+}