+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Please note this changelog affects 
 this package and not the Oxxa API.
 
+## [2.4.0]
+
+### Added
+
+- Allow additional parameters for the ssl products endpoint.
+
+### Changed
+
+- Moved pricing to a `Price` object.
+
 ## [2.3.0]
 
 ### Added